// apps/desktop/src/types/api/StockPortfolio.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StockPosition } from "./StockPosition";

export type StockPortfolio = { invested: number, marketValue: number | null, unrealisedProfitLoss: number | null, realisedProfitLoss: number, positions: Array<StockPosition>, };